I have fixed some mistakes in the bd , backend , frontend and ive incoporate new atributes

+++ StoreCollection/src/main/java/com/proyecto/StoreCollection/controller/AtributoController.java
@@ -38,13 +38,13 @@ public class AtributoController {
     }
 
     @PutMapping("/api/owner/atributos/{id}")
-    public ResponseEntity<AtributoResponse> actualizar(@PathVariable Long id,
+    public ResponseEntity<AtributoResponse> actualizar(@PathVariable Integer id,
                                                        @Valid @RequestBody AtributoRequest request) {
         return ResponseEntity.ok(service.save(request, id));
     }
 
     @DeleteMapping("/api/owner/atributos/{id}")
-    public ResponseEntity<Void> eliminar(@PathVariable Long id) {
+    public ResponseEntity<Void> eliminar(@PathVariable Integer id) {
         service.deleteById(id);
         return ResponseEntity.noContent().build();
     }
