+++ StoreCollection/src/main/java/com/proyecto/StoreCollection/dto/request/AtributoValorRequest.java
@@ -7,5 +7,5 @@ import lombok.Data;
 public class AtributoValorRequest {
     @NotBlank
     private String valor;
-    private Long atributoId;
+    private Integer atributoId;
 }
