+++ StoreCollection/src/main/java/com/proyecto/StoreCollection/dto/response/ProductoResponse.java
@@ -8,11 +8,11 @@ import lombok.NoArgsConstructor;
 @Data
 @AllArgsConstructor
 public class ProductoResponse {
-    private Long id;
+    private Integer id;
     private String nombre;
     private String slug;
-    private Long categoriaId;
+    private Integer categoriaId;
     private String categoriaNombre;   
-    private Long tiendaId;
+    private Integer tiendaId;
 
 }
