+++ StoreCollection/src/main/java/com/proyecto/StoreCollection/service/PlanService.java
@@ -7,8 +7,8 @@ import org.springframework.data.domain.Pageable;
 
 public interface PlanService {
     Page<PlanResponse> findAll(Pageable pageable);
-    PlanResponse findById(Long id);
+    PlanResponse findById(Integer id);
     PlanResponse save(PlanRequest request);
-    PlanResponse save(PlanRequest request, Long id);
-    void deleteById(Long id);
+    PlanResponse save(PlanRequest request, Integer id);
+    void deleteById(Integer id);
 }
